Fix #19 by only building for supported ABIs
Currently the connected devices SDK only is built armeabi-v7a.
Android's native library extraction looks in order of precedence for matching
abis and stops at the first one it finds in the app that can be loaded. In this
case that was arm64. This meant that the libcdp_sdk.so file located in /lib/armeabi-v7a
wasn't being found. The fix here is to remove the arm64 folder.