Merged PR 13893: Update iOS SDK Sample for new init flow
This change updates the iOS sample to use the new v1.0.0 init flow. This is a fair bit of complicated setup so it is recommended that all app developers try to follow the model laid out by this sample pretty closely. It shows how to synchronize accounts with a token library at startup, perform per account initialization as fast as possible to let incoming notifications be quickly processed and has a reusable manager object that is not tightly coupled to UI.