Merged PR 14205: Updating Android GraphNotificationsSample for improved patterns
Updating the GraphNotificationsSample to follow initialization patterns of the sdksample. It is fairly complicated setup, so it is recommended that all app developers try to follow the model laid out by this sample pretty closely. It shows how to synchronize accounts with a token library at startup, perform per account initialization as fast as possible to let incoming notifications be quickly processed and has a reusable manager object that is not tightly coupled to UI.