Merged PR 15306: Initial Check-in for App Service Code to demonstrate OBO Flow for Graph Notifications
This is the sample for an AppService that can create a notification for the user using OBO flow.
Ideally, user credentials are sent from client app to the app service, but since that pipeline is not available and not in scope for this sample, the user signs directly into the app service to provide their credentials. This part of the code is directly labeled as "client side code" in the js files to clarify how code should be structured. The main sample is the OBO flow.

**DISABLE_SECRET_SCANNING** - there is a key "AppServiceSecret" with no value. This triggers the secret scanning even though there are no secrets in the code.